apps/api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
pydantic-settings==2.6.0
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
python-jose==3.3.0
passlib==1.7.4
apscheduler==3.10.4
